Semi-polished end version.
Lot of usability changes, the script now use /etc/portage/package file
by default, the use has to specify the type using -t.

Moved the type detection to the argument parser and out of the Package
class.

The script will no longer throw exceptions when the file does not exist,
and it will simply try to create an empty one, perhaps in the future it
will destroy empty files, permissions were left untouched and python
throws the appropriate exceptions for them.

Added the _atom_rule() method that takes an atom returns its rule in a
one line string format, this is used alot when writing rules to files
are just printing the status to the screen.

The convert method now handles creating the new file/folder,
save_rules() only writes to them.

The script will now display the rule for the specified atom if no flags
were given, with the exception for mask/unmask packages since they do no
require flags.