Prompt the user before creating an empty package file.
During initialization the script will check if the package file for the
specified type exists, if not it will prompt the user to create an empty
one, otherwise exit since there is nothing to be done.

the -s flag will also print a message saying that a file has no rules in
it instead of printing nothing.